requirements.txt
# python v.3.13

transformers==4.57.3
faiss-cpu==1.13.1
pillow==12.0.0
tqdm==4.67.1
torch>=2.7.0
scikit-learn==1.7.2
scipy==1.16.3

huggingface-hub>=0.34.0,<1.0
sentence-transformers==5.1.2
langchain==1.1.3
langchain-community==0.4.1
langchain-core==1.1.2
langchain-huggingface==1.1.0
langchain-mistralai==0.2.6
langchain-text-splitters==1.0.0

numpy==2.3.5
pandas==2.3.3

pydantic==2.12.5
aiogram==3.23.0
python-dotenv==1.2.1
mistralai==1.9.11